requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
dj-database-url==0.4.1
Django==2.1.2
django-pylibmc==0.6.1
djangorestframework==3.8.2
elo==0.1.1
gunicorn==19.6.0
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
psycopg2-binary==2.7.5
pycrypto==2.6.1
pygobject==3.26.1
pylibmc==1.5.1
pytz==2018.5
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
whitenoise==3.2.2